moving api handler out of struct
