+++ gateway/api.go
@@ -31,12 +31,9 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"io/ioutil"
 	"net/http"
 	"net/url"
-	"os"
-	"path/filepath"
 	"strconv"
 	"strings"
 	"sync"
@@ -46,7 +43,6 @@ import (
 	"github.com/gorilla/mux"
 	uuid "github.com/satori/go.uuid"
 	"golang.org/x/crypto/bcrypt"
-	jsonpatch "gopkg.in/evanphx/json-patch.v4"
 
 	"github.com/TykTechnologies/tyk/apidef"
 	"github.com/TykTechnologies/tyk/config"
@@ -674,172 +670,6 @@ func handleDeleteHashedKey(keyName, apiID string) (interface{}, int) {
 	return statusObj, http.StatusOK
 }
 
-func handleGetAPIList() (interface{}, int) {
-	apisMu.RLock()
-	defer apisMu.RUnlock()
-	apiIDList := make([]*apidef.APIDefinition, len(apisByID))
-	c := 0
-	for _, apiSpec := range apisByID {
-		apiIDList[c] = apiSpec.APIDefinition
-		c++
-	}
-	return apiIDList, http.StatusOK
-}
-
-func handleGetAPI(apiID string) (interface{}, int) {
-	if spec := getApiSpec(apiID); spec != nil {
-		return spec.APIDefinition, http.StatusOK
-	}
-
-	log.WithFields(logrus.Fields{
-		"prefix": "api",
-		"apiID":  apiID,
-	}).Error("API doesn't exist.")
-	return apiError("API not found"), http.StatusNotFound
-}
-
-func handleAddOrUpdateApi(apiID string, r *http.Request) (interface{}, int) {
-	if config.Global().UseDBAppConfigs {
-		log.Error("Rejected new API Definition due to UseDBAppConfigs = true")
-		return apiError("Due to enabled use_db_app_configs, please use the Dashboard API"), http.StatusInternalServerError
-	}
-
-	newDef := &apidef.APIDefinition{}
-	if err := json.NewDecoder(r.Body).Decode(newDef); err != nil {
-		log.Error("Couldn't decode new API Definition object: ", err)
-		return apiError("Request malformed"), http.StatusBadRequest
-	}
-
-	if apiID != "" && newDef.APIID != apiID {
-		log.Error("PUT operation on different APIIDs")
-		return apiError("Request APIID does not match that in Definition! For Updtae operations these must match."), http.StatusBadRequest
-	}
-
-	// Create a filename
-	defFilePath := filepath.Join(config.Global().AppPath, newDef.APIID+".json")
-
-	// If it exists, delete it
-	if _, err := os.Stat(defFilePath); err == nil {
-		log.Warning("API Definition with this ID already exists, deleting file...")
-		os.Remove(defFilePath)
-	}
-
-	// unmarshal the object into the file
-	asByte, err := json.MarshalIndent(newDef, "", "  ")
-	if err != nil {
-		log.Error("Marshalling of API Definition failed: ", err)
-		return apiError("Marshalling failed"), http.StatusInternalServerError
-	}
-
-	if err := ioutil.WriteFile(defFilePath, asByte, 0644); err != nil {
-		log.Error("Failed to create file! - ", err)
-		return apiError("File object creation failed, write error"), http.StatusInternalServerError
-	}
-
-	action := "modified"
-	if r.Method == "POST" {
-		action = "added"
-	}
-
-	response := apiModifyKeySuccess{
-		Key:    newDef.APIID,
-		Status: "ok",
-		Action: action,
-	}
-
-	return response, http.StatusOK
-}
-
-// TODO - This is WiP. Requires cleaning up & removing duplicate code from handleAddOrUpdateApi
-func handlePatchAPI(apiID string, data []byte) (interface{}, int) {
-	if config.Global().UseDBAppConfigs {
-		log.Error("Rejected new API Definition due to UseDBAppConfigs = true")
-		return apiError("Due to enabled use_db_app_configs, please use the Dashboard API"), http.StatusBadRequest
-	}
-
-	spec := getApiSpec(apiID)
-	if spec == nil {
-		return "api does not exist", http.StatusNotFound
-	}
-
-	// unmarshal the object into the file
-	apiDefBytes, err := json.Marshal(spec.APIDefinition)
-	if err != nil {
-		log.Error("Marshalling of API Definition failed: ", err)
-		return apiError("Marshalling failed"), http.StatusInternalServerError
-	}
-
-	patch, err := jsonpatch.DecodePatch(data)
-	if err != nil {
-		return apiError(fmt.Sprintf("decodePatch failed: %v", err)), http.StatusBadRequest
-	}
-
-	modified, err := patch.Apply(apiDefBytes)
-	if err != nil {
-		return apiError(fmt.Sprintf("patchApply failed: %v", err)), http.StatusBadRequest
-	}
-
-	var dest apidef.APIDefinition
-	if err := json.Unmarshal(modified, &dest); err != nil {
-		return apiError("unable to re-encode to apidef"), http.StatusInternalServerError
-	}
-
-	if apiID != dest.APIID {
-		log.Error("Patch operation on different APIIDs")
-		return apiError("Request APIID does not match that in Definition! For Updtae operations these must match."), http.StatusBadRequest
-	}
-
-	// Create a filename
-	defFilePath := filepath.Join(config.Global().AppPath, dest.APIID+".json")
-
-	// If it exists, delete it
-	if _, err := os.Stat(defFilePath); err == nil {
-		log.Warning("API Definition with this ID already exists, deleting file...")
-		os.Remove(defFilePath)
-	}
-
-	// unmarshal the object into the file
-	asByte, err := json.MarshalIndent(dest, "", "  ")
-	if err != nil {
-		log.Error("Marshalling of API Definition failed: ", err)
-		return apiError("Marshalling failed"), http.StatusInternalServerError
-	}
-
-	if err := ioutil.WriteFile(defFilePath, asByte, 0644); err != nil {
-		log.Error("Failed to create file! - ", err)
-		return apiError("File object creation failed, write error"), http.StatusInternalServerError
-	}
-
-	response := apiModifyKeySuccess{
-		Key:    dest.APIID,
-		Status: "ok",
-		Action: "patch",
-	}
-
-	return response, http.StatusOK
-}
-
-func handleDeleteAPI(apiID string) (interface{}, int) {
-	// Generate a filename
-	defFilePath := filepath.Join(config.Global().AppPath, apiID+".json")
-
-	// If it exists, delete it
-	if _, err := os.Stat(defFilePath); err != nil {
-		log.Warning("File does not exist! ", err)
-		return apiError("Delete failed"), http.StatusInternalServerError
-	}
-
-	os.Remove(defFilePath)
-
-	response := apiModifyKeySuccess{
-		Key:    apiID,
-		Status: "ok",
-		Action: "deleted",
-	}
-
-	return response, http.StatusOK
-}
-
 func keyHandler(w http.ResponseWriter, r *http.Request) {
 	keyName := mux.Vars(r)["keyName"]
 	apiID := r.URL.Query().Get("api_id")
@@ -980,148 +810,6 @@ func handleUpdateHashedKey(keyName, apiID, policyId string) (interface{}, int) {
 		"status": "ok",
 	}).Info("Updated hashed key.")
 
-	return statusObj, http.StatusOK
-}
-
-func handleOrgAddOrUpdate(keyName string, r *http.Request) (interface{}, int) {
-	newSession := new(user.SessionState)
-
-	if err := json.NewDecoder(r.Body).Decode(newSession); err != nil {
-		log.Error("Couldn't decode new session object: ", err)
-		return apiError("Request malformed"), http.StatusBadRequest
-	}
-	// Update our session object (create it)
-
-	spec := getSpecForOrg(keyName)
-	var sessionManager SessionHandler
-
-	if spec == nil {
-		log.Warning("Couldn't find org session store in active API list")
-		if config.Global().SupressDefaultOrgStore {
-			return apiError("No such organisation found in Active API list"), http.StatusNotFound
-		}
-		sessionManager = &DefaultOrgStore
-	} else {
-		sessionManager = spec.OrgSessionManager
-	}
-
-	if r.URL.Query().Get("reset_quota") == "1" {
-		sessionManager.ResetQuota(keyName, newSession, false)
-		newSession.QuotaRenews = time.Now().Unix() + newSession.QuotaRenewalRate
-		rawKey := QuotaKeyPrefix + storage.HashKey(keyName)
-
-		// manage quotas separately
-		DefaultQuotaStore.RemoveSession(rawKey, false)
-	}
-
-	err := sessionManager.UpdateSession(keyName, newSession, 0, false)
-	if err != nil {
-		return apiError("Error writing to key store " + err.Error()), http.StatusInternalServerError
-	}
-
-	// identify that spec has org session
-	if spec != nil {
-		spec.Lock()
-		spec.OrgHasNoSession = false
-		spec.Unlock()
-	}
-
-	log.WithFields(logrus.Fields{
-		"prefix": "api",
-		"org":    keyName,
-		"status": "ok",
-	}).Info("New organization key added or updated.")
-
-	action := "modified"
-	if r.Method == "POST" {
-		action = "added"
-	}
-
-	response := apiModifyKeySuccess{
-		Key:    keyName,
-		Status: "ok",
-		Action: action,
-	}
-
-	return response, http.StatusOK
-}
-
-func handleGetOrgDetail(orgID string) (interface{}, int) {
-	spec := getSpecForOrg(orgID)
-	if spec == nil {
-		return apiError("Org not found"), http.StatusNotFound
-	}
-
-	session, ok := spec.OrgSessionManager.SessionDetail(orgID, false)
-	if !ok {
-		log.WithFields(logrus.Fields{
-			"prefix": "api",
-			"org":    orgID,
-			"status": "fail",
-			"err":    "not found",
-		}).Error("Failed retrieval of record for ORG ID.")
-		return apiError("Org not found"), http.StatusNotFound
-	}
-	log.WithFields(logrus.Fields{
-		"prefix": "api",
-		"org":    orgID,
-		"status": "ok",
-	}).Info("Retrieved record for ORG ID.")
-	return session, http.StatusOK
-}
-
-func handleGetAllOrgKeys(filter string) (interface{}, int) {
-	spec := getSpecForOrg("")
-	if spec == nil {
-		return apiError("ORG not found"), http.StatusNotFound
-	}
-
-	sessions := spec.OrgSessionManager.Sessions(filter)
-	fixed_sessions := make([]string, 0)
-	for _, s := range sessions {
-		if !strings.HasPrefix(s, QuotaKeyPrefix) && !strings.HasPrefix(s, RateLimitKeyPrefix) {
-			fixed_sessions = append(fixed_sessions, s)
-		}
-	}
-	sessionsObj := apiAllKeys{fixed_sessions}
-	return sessionsObj, http.StatusOK
-}
-
-func handleDeleteOrgKey(orgID string) (interface{}, int) {
-	spec := getSpecForOrg(orgID)
-	if spec == nil {
-		log.WithFields(logrus.Fields{
-			"prefix": "api",
-			"key":    orgID,
-			"status": "fail",
-			"err":    "not found",
-		}).Error("Failed to delete org key.")
-
-		return apiError("Org not found"), http.StatusNotFound
-	}
-
-	if !spec.OrgSessionManager.RemoveSession(orgID, false) {
-		return apiError("Failed to remove the key"), http.StatusBadRequest
-	}
-
-	log.WithFields(logrus.Fields{
-		"prefix": "api",
-		"key":    orgID,
-		"status": "ok",
-	}).Info("Org key deleted.")
-
-	// identify that spec has no org session
-	if spec != nil {
-		spec.Lock()
-		spec.OrgHasNoSession = true
-		spec.Unlock()
-	}
-
-	statusObj := apiModifyKeySuccess{
-		Key:    orgID,
-		Status: "ok",
-		Action: "deleted",
-	}
 	return statusObj, http.StatusOK
 }
